sploits/memos/requirements.txt
requests==2.31.0
Pillow==10.1.0
